[#63] 스타일 : category -> breadCategory 명칭 수정
